Move metadata dictionary to Microscope object.
You can now change the patient id, slide id, sample id, operator
id tags by accessing the local Microscope object's "metadata"
dictionary. You do this before calling camera.capture, so that
the metadata is written as exif tags to the image. Previously,
this code only existed for taking an image manually by pressing
"j". Keeping the metadata as a field of the Microscope object
seems to be the simplest solution at the time.